ci: split oversized Rust modules
