+++ crates/bentodesk-shell/src/shell/pointer_up_drag.rs
@@ -458,540 +458,3 @@ pub(super) fn handle_active_pointer_drag(
 
     false
 }
-
-/// Clamp a live Zone drag to the Main HWND's logical viewport.
-///
-/// Device-monitor bounds are intentionally accepted only as regression-proof
-/// context: Main-client logical DIPs are the persistence coordinate space and
-/// must not roam into a larger physical/virtual-screen union.
-pub(super) fn clamp_zone_drag_to_logical_viewport(
-    zones: &ZoneList,
-    id: ZoneId,
-    x: i32,
-    y: i32,
-    viewport: bentodesk_style::Size,
-    _device_monitors: &[bentodesk_platform::MonitorInfo],
-) -> (i32, i32) {
-    let Some(zone) = zones.get(id) else {
-        return (x, y);
-    };
-    let (_, _, width, height) =
-        bentodesk_app::zone_gesture_geometry::zone_drag_capsule_rect(zones, zone);
-    bentodesk_app::zone_pill_geometry::clamp_capsule_origin_to_viewport(
-        x, y, width, height, viewport,
-    )
-}
-
-#[cfg(test)]
-#[allow(clippy::items_after_test_module)]
-mod viewport_drag_tests {
-    use super::*;
-
-    #[test]
-    fn logical_viewport_clamp_covers_dpi_taskbar_edges_and_capsule_footprints() {
-        use bentodesk_platform::RectI32;
-        use bentodesk_style::Size;
-
-        let viewport_cases = [
-            (
-                "100%-bottom",
-                96,
-                Size {
-                    width: 1_920.75,
-                    height: 1_040.75,
-                },
-                RectI32 {
-                    left: 0,
-                    top: 0,
-                    right: 3_840,
-                    bottom: 2_160,
-                },
-                RectI32 {
-                    left: 0,
-                    top: 0,
-                    right: 3_840,
-                    bottom: 2_080,
-                },
-            ),
-            (
-                "125%-top",
-                120,
-                Size {
-                    width: 1_920.75,
-                    height: 1_040.75,
-                },
-                RectI32 {
-                    left: 0,
-                    top: 0,
-                    right: 2_400,
-                    bottom: 1_350,
-                },
-                RectI32 {
-                    left: 0,
-                    top: 50,
-                    right: 2_400,
-                    bottom: 1_350,
-                },
-            ),
-            (
-                "150%-left",
-                144,
-                Size {
-                    width: 1_872.75,
-                    height: 1_080.75,
-                },
-                RectI32 {
-                    left: 0,
-                    top: 0,
-                    right: 2_880,
-                    bottom: 1_620,
-                },
-                RectI32 {
-                    left: 72,
-                    top: 0,
-                    right: 2_880,
-                    bottom: 1_620,
-                },
-            ),
-            (
-                "200%-right",
-                192,
-                Size {
-                    width: 1_872.75,
-                    height: 1_080.75,
-                },
-                RectI32 {
-                    left: 0,
-                    top: 0,
-                    right: 3_840,
-                    bottom: 2_160,
-                },
-                RectI32 {
-                    left: 0,
-                    top: 0,
-                    right: 3_744,
-                    bottom: 2_160,
-                },
-            ),
-        ];
-        let footprints = [
-            ("small-pill", "small", "pill", false),
-            ("medium-pill", "medium", "pill", false),
-            ("large-pill", "large", "pill", false),
-            ("small-circle", "small", "circle", false),
-            ("stack", "medium", "pill", true),
-        ];
-
-        for (viewport_name, dpi, viewport, rect_screen, rect_work) in viewport_cases {
-            let device_monitor = bentodesk_platform::MonitorInfo {
-                hmonitor: core::ptr::null_mut(),
-                rect_screen,
-                rect_work,
-                is_primary: true,
-            };
-            assert!(
-                rect_work.right - rect_work.left > viewport.width.floor() as i32
-                    || rect_work.bottom - rect_work.top > viewport.height.floor() as i32,
-                "fixture must expose inflated device bounds: {viewport_name} dpi={dpi}"
-            );
-
-            for (footprint_name, size, shape, stack) in footprints {
-                let mut zones = ZoneList::new();
-                let mut anchor = Zone::new(ZoneId(1), "zone", 10, 10, 480, 432);
-                anchor.set_capsule(size, shape);
-                zones.add(anchor);
-                if stack {
-                    zones.add(Zone::new(ZoneId(2), "child", 20, 20, 320, 240));
-                    assert!(zones.stack(ZoneId(1), ZoneId(2)));
-                }
-                let zone = zones.get(ZoneId(1)).expect("zone");
-                let (_, _, capsule_width, capsule_height) =
-                    bentodesk_app::zone_gesture_geometry::zone_drag_capsule_rect(&zones, zone);
-                let expected = (
-                    (viewport.width.floor() as i32 - capsule_width).max(0),
-                    (viewport.height.floor() as i32 - capsule_height).max(0),
-                );
-
-                assert_eq!(
-                    clamp_zone_drag_to_logical_viewport(
-                        &zones,
-                        ZoneId(1),
-                        i32::MAX / 4,
-                        i32::MAX / 4,
-                        viewport,
-                        &[device_monitor],
-                    ),
-                    expected,
-                    "{viewport_name} dpi={dpi} footprint={footprint_name}"
-                );
-                assert_eq!(
-                    clamp_zone_drag_to_logical_viewport(
-                        &zones,
-                        ZoneId(1),
-                        -500,
-                        -500,
-                        viewport,
-                        &[device_monitor],
-                    ),
-                    (0, 0),
-                    "negative clamp: {viewport_name} footprint={footprint_name}"
-                );
-            }
-        }
-    }
-}
-
-pub(super) fn item_external_drag_modifier_down() -> bool {
-    // SAFETY: GetAsyncKeyState is a read-only user32 query and is valid from
-    // the UI thread while handling mouse input. Ctrl follows the Explorer
-    // convention for "copy this file payload out to another target" and avoids
-    // stealing the default in-zone reorder / cross-zone move path.
-    unsafe { (GetAsyncKeyState(VK_CONTROL as i32) as u16 & 0x8000) != 0 }
-}
-
-pub(super) struct ItemDragOutDropTargetGuard<'a> {
-    root: &'a AppRoot,
-}
-
-impl Drop for ItemDragOutDropTargetGuard<'_> {
-    fn drop(&mut self) {
-        self.root.item_drag_out_active.set(false);
-        log_static("items: drag-out self-drop-target resumed\n");
-    }
-}
-
-pub(super) fn with_item_drag_out_guard<T>(root: &AppRoot, drag: impl FnOnce() -> T) -> T {
-    root.item_drag_out_active.set(true);
-    log_static("items: drag-out self-drop-target suspended\n");
-    let _guard = ItemDragOutDropTargetGuard { root };
-    drag()
-}
-
-pub(super) fn start_item_drag_out(root: &AppRoot, source_hwnd: HWND, request: PendingItemDragOut) {
-    if request.path.is_empty() {
-        return;
-    }
-    let hwnd_bits = source_hwnd as isize;
-    let path = request.path.to_string();
-    let leaf = item_operation_leaf(path.as_str()).to_owned();
-    log_static(format!("items: drag-out started path={path}\n").as_str());
-    set_item_operation_status(
-        root,
-        localized_current(
-            format!("正在拖出：{}", item_operation_leaf(path.as_str())),
-            format!("Dragging out: {}", item_operation_leaf(path.as_str())),
-        ),
-    );
-    let result = with_item_drag_out_guard(root, || {
-        bentodesk_backend::drag_drop::start_drag_operation_from_hwnd(
-            std::slice::from_ref(&path),
-            hwnd_bits,
-            request.copy_only,
-        )
-    });
-    match result {
-        Ok(outcome) => {
-            tracing::info!(
-                target: "bentodesk::drag_drop",
-                %path,
-                outcome = %outcome.as_str(),
-                "item drag-out completed"
-            );
-            log_static(
-                format!(
-                    "items: drag-out completed path={} outcome={}\n",
-                    path,
-                    outcome.as_str()
-                )
-                .as_str(),
-            );
-            finalize_item_drag_out(root, &request, leaf.as_str(), outcome);
-        }
-        Err(error) => {
-            tracing::warn!(
-                target: "bentodesk::drag_drop",
-                %path,
-                error = %error,
-                "item drag-out failed"
-            );
-            log_static(format!("items: drag-out failed path={path} error={error}\n").as_str());
-            set_item_operation_status(
-                root,
-                localized_current(
-                    format!("拖出失败：{leaf}：{error}"),
-                    item_drag_out_status_for_error_message(leaf.as_str(), &error.to_string()),
-                ),
-            );
-        }
-    }
-    if hwnd_bits != 0 {
-        // SAFETY: hwnd_bits came from a live process-owned HWND. Invalidating
-        // after the Shell drag loop returns updates the visible status.
-        unsafe { InvalidateRect(hwnd_bits as HWND, ptr::null(), 0) };
-    }
-}
-
-pub(super) fn finalize_item_drag_out(
-    root: &AppRoot,
-    request: &PendingItemDragOut,
-    leaf: &str,
-    outcome: bentodesk_backend::drag_drop::DragOutcome,
-) {
-    match outcome {
-        bentodesk_backend::drag_drop::DragOutcome::Copied => {
-            set_item_operation_status(
-                root,
-                localized_current(
-                    format!("已复制到外部：{leaf}"),
-                    format!("Copied out: {leaf}"),
-                ),
-            );
-        }
-        bentodesk_backend::drag_drop::DragOutcome::Moved => {
-            let source_missing = match source_missing_after_shell_move(Path::new(
-                request.path.as_str(),
-            )) {
-                Ok(missing) => missing,
-                Err(error) => {
-                    set_item_operation_status(
-                        root,
-                        localized_current(
-                            format!("无法确认移出结果，已保留：{leaf}"),
-                            format!("Move could not be verified; kept: {leaf}"),
-                        ),
-                    );
-                    log_static(
-                        format!(
-                            "items: drag-out model-kept source-metadata-error zone={} item={} path={} error={}\n",
-                            request.zone_id.0, request.item_id.0, request.path, error
-                        )
-                        .as_str(),
-                    );
-                    return;
-                }
-            };
-            if !source_missing {
-                // A target can report MOVE while expecting the source to delete
-                // the original (an unoptimised move). BentoDesk never deletes
-                // user bytes during drag-out, so keep the model whenever the
-                // source still exists instead of manufacturing data loss.
-                set_item_operation_status(
-                    root,
-                    localized_current(
-                        format!("已复制到外部：{leaf}"),
-                        format!("Copied out: {leaf}"),
-                    ),
-                );
-                log_static(
-                    format!(
-                        "items: drag-out model-kept source-still-exists zone={} item={} path={}\n",
-                        request.zone_id.0, request.item_id.0, request.path
-                    )
-                    .as_str(),
-                );
-                return;
-            }
-            // The Shell has already completed the MOVE represented by
-            // `DROPEFFECT_MOVE`. For stealth-backed items that means the hidden
-            // source path no longer exists: routing through ordinary RemoveItem
-            // would try to restore that already-moved file, fail, and leave a
-            // ghost card in the Zone. This completion path owns model cleanup
-            // only; filesystem ownership has transferred to the drop target.
-            remove_item_model_after_shell_move(root, request.zone_id, request.item_id);
-            let removed = root
-                .app
-                .borrow()
-                .zones
-                .item(request.zone_id, request.item_id)
-                .is_none();
-            if removed {
-                flush_dirty_zones(root);
-                set_item_operation_status(
-                    root,
-                    localized_current(format!("已移出：{leaf}"), format!("Moved out: {leaf}")),
-                );
-                log_static(
-                    format!(
-                        "items: drag-out model-removed zone={} item={} path={}\n",
-                        request.zone_id.0, request.item_id.0, request.path
-                    )
-                    .as_str(),
-                );
-            } else {
-                log_static(
-                    format!(
-                        "items: drag-out model-kept zone={} item={} path={}\n",
-                        request.zone_id.0, request.item_id.0, request.path
-                    )
-                    .as_str(),
-                );
-            }
-        }
-        bentodesk_backend::drag_drop::DragOutcome::Dropped => {
-            // A completed drop without one exact COPY/MOVE effect is not
-            // evidence that the source left disk. Keep the Zone model.
-            set_item_operation_status(
-                root,
-                localized_current(
-                    format!("已拖出：{leaf}"),
-                    item_drag_out_status_for_outcome(leaf, outcome),
-                ),
-            );
-        }
-        bentodesk_backend::drag_drop::DragOutcome::Cancelled => {
-            set_item_operation_status(
-                root,
-                localized_current(
-                    format!("已取消拖出：{leaf}"),
-                    item_drag_out_status_for_outcome(leaf, outcome),
-                ),
-            );
-        }
-    }
-}
-
-pub(super) fn source_missing_after_shell_move(path: &Path) -> std::io::Result<bool> {
-    source_missing_from_metadata(std::fs::symlink_metadata(path))
-}
-
-pub(super) fn source_missing_from_metadata(
-    metadata: std::io::Result<std::fs::Metadata>,
-) -> std::io::Result<bool> {
-    match metadata {
-        Ok(_) => Ok(false),
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => Ok(true),
-        Err(error) => Err(error),
-    }
-}
-
-pub(super) fn remove_item_model_after_shell_move(
-    root: &AppRoot,
-    zone_id: ZoneId,
-    item_id: bentodesk_zone::ZoneItemId,
-) {
-    let removed_item = {
-        let mut app = root.app.borrow_mut();
-        let item = app.zones.item(zone_id, item_id).cloned();
-        if app.zones.remove_item(zone_id, item_id) {
-            app.mark_dirty();
-            item
-        } else {
-            None
-        }
-    };
-    let Some(item) = removed_item else {
-        return;
-    };
-    let (Some(original), Some(hidden)) =
-        (item.original_path.as_deref(), item.hidden_path.as_deref())
-    else {
-        return;
-    };
-    if Path::new(hidden).exists() {
-        return;
-    }
-    let Some(config) = stealth_config_for_source(root, original) else {
-        return;
-    };
-    let result = bentodesk_backend::stealth::hidden_dir_for(&config)
-        .and_then(|dir| bentodesk_backend::stealth::manifest_remove(&dir, original));
-    if let Err(error) = result {
-        tracing::warn!(
-            target: "bentodesk::stealth",
-            original,
-            hidden,
-            %error,
-            "drag-out completed but stale recovery manifest entry could not be removed"
-        );
-    }
-}
-
-#[cfg(test)]
-pub(super) fn start_item_drag_out_with<StartDrag>(
-    root: &AppRoot,
-    path: String,
-    start_drag: StartDrag,
-) -> bool
-where
-    StartDrag: FnOnce(
-        &[String],
-    ) -> Result<
-        bentodesk_backend::drag_drop::DragOutcome,
-        bentodesk_backend::drag_drop::DragDropError,
-    >,
-{
-    if path.is_empty() {
-        return false;
-    }
-    let leaf = item_operation_leaf(path.as_str()).to_owned();
-    log_static(format!("items: drag-out started path={path}\n").as_str());
-    match with_item_drag_out_guard(root, || start_drag(std::slice::from_ref(&path))) {
-        Ok(outcome) => {
-            tracing::info!(
-                target: "bentodesk::drag_drop",
-                %path,
-                outcome = %outcome.as_str(),
-                "item drag-out completed"
-            );
-            log_static(
-                format!(
-                    "items: drag-out completed path={} outcome={}\n",
-                    path,
-                    outcome.as_str()
-                )
-                .as_str(),
-            );
-            set_item_operation_status(
-                root,
-                SmolStr::new(item_drag_out_status_for_outcome(leaf.as_str(), outcome)),
-            );
-            true
-        }
-        Err(e) => {
-            tracing::warn!(
-                target: "bentodesk::drag_drop",
-                %path,
-                error = %e,
-                "item drag-out failed"
-            );
-            log_static(format!("items: drag-out failed path={path} error={e}\n").as_str());
-            set_item_operation_status(
-                root,
-                SmolStr::new(item_drag_out_status_for_error(leaf.as_str(), &e)),
-            );
-            true
-        }
-    }
-}
-
-pub(super) fn item_drag_out_status_for_outcome(
-    leaf: &str,
-    outcome: bentodesk_backend::drag_drop::DragOutcome,
-) -> String {
-    match outcome {
-        bentodesk_backend::drag_drop::DragOutcome::Copied => {
-            format!("Copied out: {leaf}")
-        }
-        bentodesk_backend::drag_drop::DragOutcome::Moved => {
-            format!("Moved out: {leaf}")
-        }
-        bentodesk_backend::drag_drop::DragOutcome::Dropped => {
-            format!("Dragged out: {leaf}")
-        }
-        bentodesk_backend::drag_drop::DragOutcome::Cancelled => {
-            format!("Drag out cancelled: {leaf}")
-        }
-    }
-}
-
-#[cfg(test)]
-pub(super) fn item_drag_out_status_for_error(
-    leaf: &str,
-    error: &bentodesk_backend::drag_drop::DragDropError,
-) -> String {
-    format!("Drag out failed: {leaf}: {error}")
-}
-
-pub(super) fn item_drag_out_status_for_error_message(leaf: &str, error: &str) -> String {
-    format!("Drag out failed: {leaf}: {error}")
-}
